model traits refactoring

+++ app/Models/Traits/HasPosts.php
@@ -8,18 +8,20 @@ use Illuminate\Database\Eloquent\Relations\MorphMany;
 
 trait HasPosts
 {
-    public function posts():MorphMany
-    {
-        return $this->morphMany(Post::class, 'pageable');
-    }
-
     public static function bootHasPosts()
     {
         static::deleting(function(Model $pageable){
             assertInTransaction();
-            $pageable->posts()->cursor()->each(function(Post $post){
-                $post->forceDelete();
-            });
+            if($pageable->forceDeleting())
+            {
+                $pageable->posts()->cursor()->each(function(Post $post){
+                    $post->forceDelete();
+                });
+            }
         });
+    }
+    public function posts():MorphMany
+    {
+        return $this->morphMany(Post::class, 'pageable');
     }
 }
